Extract voucher view tests to the separate module and move to the proper test package.
